Added teensy related docs, datasheets. Refactored Task
Stylistic corrections to Task like camelCase-ing and better variable names.
Made the static `Task::taskCount` private, now its called Task::_taskCount.
