+++ arduino/task/keywords.txt
@@ -1,0 +1,7 @@
+# Syntax highlighting
+
+# Class (KEYWORD1)
+Task	KEYWORD1
+
+# Public methods(KEYWORD2)
+computePriority	KEYWORD2
