+++ teensy/designs/audio.md
@@ -1,0 +1,35 @@
+# Teensy Audio
+
+The Teensy Audio library supports the following without any new hardware:
+
+* synthesis of tones and waveforms and o/p using DAC / PWM
+* DSP on waveforms
+	- stored in memory (also SD card)
+	- recieved as analog signal on ADC
+
+The Audio Adapter board (SGTL5000) and the Teensy share signals on the I2S lines. Teensy uses I2C to control/program SGTL5000.
+
+## Sending audio to Teensy
+Analog signals from PC `Audio Out` can be fed to the Audio board and recieved by the Teensy via I2S.
+```
+PC --headphone out--> Audio Board --I2S--> Teensy --SPI--> Cyclops
+```
+
+As the waveform is generated by the OE GUI, it can be "played" on the Teensy via I2S.
+> Since sound is stereo, 2 channels can be recieved in real time at 44kHz.
+
+`44 kHz * 16b * 2 = 1.4112 MHz`
+Tha baud rate of Serial interface is comparable ~1.152Mhz (and higher too).
+_The benefit_ of using the Audio board is that `(1)` the USART link to OE GUI would be available for low latency control, waveform updates and not be busy with high bandwidth data from GUI.
+Also, the DSP on the Teensy can be setup to perform transforms on incoming data (though the GUI would probably apply the transform before sending?).
+
+This would make it easy for the Teensy to be responsive.`(2)` I2S Rx and Tx have DMA support. When the (`8 * 32b`) FIFO is filled, DMA interrupt is triggered, which could move the `32B` into the `Source.voltage_data`._Ideally we must forward the recieved 32b to `Source.voltage_data` asap_.
+
+## Using Teensy DSP to synth waveforms
+
+We should synthesise all waveforms (@ max 44kHz) audio using the DSP. The signal data-points can be sent to Cyclops over SPI. There is significant speedup in data-point generation when using the DSP (much faster 32b floating point complex mathematics).
+
+## Audio System Design GUI
+
+The functionality of this tool can be replicated easily. I suppose we would only be using the board as an input, and not for synthesis (no mixer / mux).
+The OE GUI would be able to realise the design of the Audio System and auto-generate code for the Teensy.
